Removed the index file from the new forms store.
